[cli-gui] Move some methods
